move arclength_sampling_n to trajectory
